+++ tests/testthat/test-predict-SDMmodelCV.R
@@ -30,13 +30,6 @@ test_that("The method works with raster stack objects", {
   expect_s4_class(p, "RasterLayer")
 })
 
-test_that("The method works with raster stack objects and parallel", {
-  p <- predict(m, predictors, type = "raw", parallel = TRUE)
-  expect_length(p, predictors$bio1@ncols * predictors$bio1@nrows)
-  expect_s4_class(p, "RasterLayer")
-  expect_false(getOption("SDMtuneParallel"))
-})
-
 test_that("The output is the function applied to the k predictions", {
   train@data <- train@data[1:3, ]