+++ NEWS.md
@@ -1,4 +1,4 @@
-# SDMtune 1.1.0.9000
+# SDMtune 1.1.1
 Main changes:
 * New function `checkMaxentInstallation()` to check if Maxent is correctly configured.
 * The argument parallel is deprecated and not used anymore. This because it improved computation only for very large datasets and used the superseded package snow.